requirements.txt
fastapi>=0.95.1,<0.104.0
uvicorn[standard]>=0.22.0,<0.24.0
openai>=0.28.0,<1.4.0
requests>=2.28.2,<2.32.0
python-dotenv>=1.0.0,<1.1.0
pydantic>=2.0.0,<2.5.0
httpx>=0.24.0,<0.25.0
jinja2>=3.0.0,<4.0.0 